refactor: move color to nuclues utils
